requirements.txt
Django==5.0.2
django-graphql-jwt==0.4.0
graphene==3.3
celery[redis]==5.3.6
flower==2.0.1
psycopg2-binary==2.9.9
graphene-file-upload==1.3.0
